+++ modin/core/storage_formats/pandas/utils.py
@@ -20,7 +20,8 @@ import contextlib
 import numpy as np
 import pandas
 
-from modin.config import MinPartitionSize
+from modin.config import MinPartitionSize, NPartitions
+from math import ceil
 
 
 @contextlib.contextmanager
@@ -180,3 +181,33 @@ def get_group_names(regex: "re.Pattern") -> "List[Hashable]":
     """
     names = {v: k for k, v in regex.groupindex.items()}
     return [names.get(1 + i, i) for i in range(regex.groups)]
+
+
+def merge_partitioning(left, right, axis=1):
+    """
+    Get the number of splits across the `axis` for the two dataframes being concatenated.
+
+    Parameters
+    ----------
+    left : PandasDataframe
+    right : PandasDataframe
+    axis : int, default: 1
+
+    Returns
+    -------
+    int
+    """
+    # Avoiding circular imports from pandas query compiler
+    from modin.core.storage_formats.pandas.utils import compute_chunksize
+
+    lshape = left._row_lengths_cache if axis == 0 else left._column_widths_cache
+    rshape = right._row_lengths_cache if axis == 0 else right._column_widths_cache
+
+    if lshape is not None and rshape is not None:
+        res_shape = sum(lshape) + sum(rshape)
+        chunk_size = compute_chunksize(axis_len=res_shape, num_splits=NPartitions.get())
+        return ceil(res_shape / chunk_size)
+    else:
+        lsplits = left._partitions.shape[axis]
+        rsplits = right._partitions.shape[axis]
+        return min(lsplits + rsplits, NPartitions.get())
